Tweak landing page language
There was some discussion about our philosophy on flipping items -- and
in particular the idea of asking people to never sell an item again if
they got it from us.  Bottom line is at this point we're not trying to
overstep, but also we wanna make sure people grok that using this
resource as a way of flipping items is no good.